+++ example/simple_example.dart
@@ -4,14 +4,12 @@ import 'package:tree_state_machine/tree_builders.dart';
 //
 // State keys
 //
-class SimpleStates {
+class States {
   static const enterText = StateKey('simple_enterText');
   static const showUppercase = DataStateKey<String>('simple_showUppercase');
   static const showLowercase = StateKey('simple_showLowercase');
   static const finished = DataStateKey<String>('simple_finished');
 }
-
-typedef _S = SimpleStates;
 
 //
 // Messages
@@ -32,19 +30,20 @@ class ToLowercase {
 /// states.
 class SimpleStateTree {
   StateTreeBuilder treeBuilder() {
-    var b = StateTreeBuilder(initialChild: _S.enterText, logName: 'simple');
+    var b = StateTreeBuilder(initialChild: States.enterText, logName: 'simple');
 
-    b.state(_S.enterText, (b) {
-      b.onMessage<ToUppercase>((b) => b.goTo(_S.showUppercase, payload: (ctx) => ctx.message.text));
+    b.state(States.enterText, (b) {
+      b.onMessage<ToUppercase>(
+          (b) => b.goTo(States.showUppercase, payload: (ctx) => ctx.message.text));
     });
 
     b.dataState<String>(
-      _S.showUppercase,
+      States.showUppercase,
       InitialData.run((ctx) => (ctx.payload as String).toUpperCase()),
       (b) {
         b.onMessageValue(
           Messages.finish,
-          (b) => b.goTo(SimpleStates.finished, payload: (ctx) {
+          (b) => b.goTo(States.finished, payload: (ctx) {
             return ctx.data;
           }),
         );
@@ -52,7 +51,7 @@ class SimpleStateTree {
     );
 
     b.finalDataState<String>(
-      _S.finished,
+      States.finished,
       InitialData.run((ctx) {
         return ctx.payloadOrThrow<String>();
       }),