+++ CHANGELOG.md
@@ -1,13 +1,20 @@
-## 3.0.0
+## 3.0.0-dev.5
 - Adjust parameter order of `StateTreeBuilder.withRoot` and `StateTreeBuilder.withDataRoot`
 - Add `ValueSubject.mapValueStream`.
 - Change return type of `TreeStateMachine.lifecycle` to `ValueStream`, and remove redundant getters for specific 
   lifecycle states.
 - Change return type of `TreeStateMachine.loadFrom` to `Future<CurrentState>`.
 - Add `metadata` to `TransistionContext` and `MessageContext.goTo`.
+
+## 3.0.0-dev.4
 - Add `TreeStateMachine.currentState`.
+
+## 3.0.0-dev.3
+- Add `TreeStateMachine.isStarting`.
 - Add `TreeNodeInfo.getChildren`.
-- Add `TreeStateMachine.rootNode`.
+- Experimental: Add `TreeStateMachine.rootNode`.
+
+## 3.0.0-dev.2
 - The `initialState` parameter of the `StateTreeBuilder` constructor was renamed to `initialChild` to reduce developer
   confusion, and the error message was improved when this parameter refers to an invalid state.
 - `StateTreeDefinitionError` is thrown when validating a `StateTreeBuilder`, instead of [StateError].
@@ -24,10 +31,14 @@
 - Add `leafState`, `handlingState`, and `activeStates` to `MessageContext`.
 - Add `handlingState` to `TransitionContext`.
 - Rename `MessageContext.appData` -> - `MessageContext.metadata`.
+
+## 3.0.0-dev.1
 - Upgrade to Dart3 SDK
 - Add `DataStateKey` to emphasize association between a data state and its state data type.
 - Add `TreeStateMachine.startWith` to enable starting a state machine with specific initial values.
   for data states.
+
+
 
 ## 2.4.0
 - Fix issue with `MessageHandlerWhenBuilder` not evaluating multiple conditions correctly.